requirements_migrations.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
boto3==1.34.115
botocore==1.34.115
Brotli==1.1.0
bs4==0.0.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dash==2.18.1
dash-bootstrap-components==1.3.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_colorscales==0.0.4
dash_daq==0.5.0
daphne==4.1.2
Django==4.2.13
django-crispy-forms==1.14.0
django-froala-editor==4.1.4
django-revproxy==0.12.0
django-s3direct==2.0.3
django-social-share==2.3.0
dtale==3.14.1
et-xmlfile==1.1.0
Flask==2.2.5
Flask-Compress==1.15
flask-ngrok==0.0.25
fonttools==4.53.1
future==1.0.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kaleido==0.2.1
kiwisolver==1.4.7
lz4==4.3.3
MarkupSafe==2.1.5
matplotlib==3.9.2
missingno==0.5.2
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
patsy==0.5.6
Pillow==10.0.0
plotly==5.24.1
psycopg==3.2.1
psycopg2-binary==2.8.6
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-json-logger==2.0.1
pytz==2024.2
requests==2.32.3
retrying==1.3.4
s3transfer==0.10.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
squarify==0.4.4
statsmodels==0.14.3
strsimpy==0.2.1
tenacity==9.0.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.0.4
xarray==2024.9.0
xlrd==2.0.1
zipp==3.20.2
zstandard==0.23.0
channels==4.1.0
django-apscheduler==0.6.0
django-phonenumber-field==7.3.0
django4-background-tasks==1.2.9
xgboost==1.4.2
lightgbm==3.2.1
phonenumberslite==8.13.35
django-crum==0.7.9
stripe==8.5.0
anthropic==0.26.1
openai==1.40.0
google-search-results==2.4.2
htmldate==1.8.1
aiohttp==3.9.5
aioboto3==13.1.0
tiktoken==0.6.0
google-api-python-client==2.50.0
pyspellchecker==0.7.1
json_log_formatter==0.3.1